refactor(change-review): enforce feature dependency direction
