+++ src/features/change-review/renderer/index.ts
@@ -15,9 +15,7 @@ export {
   buildReviewStats,
   buildTaskChangesEmptyStatePresentation,
   buildWatchedReviewFilePathsKey,
-  canAcceptAllReviewFiles,
   findActiveReviewFile,
-  getRejectablePendingReviewFiles,
   isTaskChangeSetV2,
   resolveReviewFileLabel,
   shouldShowTaskScopeBanner,
